+++ src/fido/HashTable.cpp
@@ -5,7 +5,7 @@ template <typename D>
 bool HashTable<D>::add(const D & data)
 {
   // add the entry to the table
-  int index = hash(data);
+  int index = static_cast<int>(hash(data));
 
   if (! table[index].empty() )
     {
@@ -29,7 +29,7 @@ bool HashTable<D>::add(const D & data)
 template <typename D>
 int HashTable<D>::lookup(const D & data) const
 {
-  int index = hash(data);
+  int index = static_cast<int>(hash(data));
   
   return searchList( table[index], data );
 }
@@ -53,7 +53,7 @@ int HashTable<D>::searchList(const list<Node> & li, const D & data) const
 template <typename D>
 unsigned int HashTable<D>::hash(const D & data) const
 {
-  return defined_hash(data) % table.size();
+  return static_cast<unsigned int>(defined_hash(data) % table.size());
 }
 
 template <typename D>
